+++ graal-js/src/com.oracle.truffle.js.test/src/com/oracle/truffle/js/test/polyglot/ForeignConsolePrintTest.java
@@ -75,9 +75,9 @@ public final class ForeignConsolePrintTest {
     public void testForeignArray() {
         final String script = "(function (a) { return '' + a; })";
         final Value fun = ctx.eval(JavaScriptLanguage.ID, script);
-        Value res = fun.execute(new ArrayTruffleObject(new int[]{0, 1, 2, 3, 4}));
+        Value res = fun.execute(new ArrayTruffleObject(new Object[]{0, 1, 2, 3, 4}));
         String sRes = res.asString();
-        assertEquals("(5)[0, 1, 2, 3, 4]", sRes);
+        assertEquals("0,1,2,3,4", sRes);
     }
 
     @Test
@@ -112,15 +112,32 @@ public final class ForeignConsolePrintTest {
 
         Value res = fun.execute(map);
         String sRes = res.asString();
-        assertEquals("{x: 42, y: \"foo\"}", sRes);
+        assertEquals("[object Object]", sRes);
+    }
+
+    @Test
+    public void testForeignArrayWithObjects() {
+        final String script = "(function (a) { return '' + a; })";
+        final Value fun = ctx.eval(JavaScriptLanguage.ID, script);
+
+        final ForeignTestMap map = new ForeignTestMap();
+        map.getContainer().put("x", 42);
+        map.getContainer().put("y", "foo");
+        ArrayTruffleObject arr = new ArrayTruffleObject(new Object[]{0, "string", map, 3, new ForeignTestFunction("test", (arg) -> {
+            return arg[0] + ", " + arg[1];
+        })});
+
+        Value res = fun.execute(arr);
+        String sRes = res.asString();
+        assertEquals("0,string,[object Object],3,function test() { [native code] }", sRes);
     }
 
     @ExportLibrary(InteropLibrary.class)
     static final class ArrayTruffleObject implements TruffleObject {
 
-        private final int[] array;
+        private final Object[] array;
 
-        ArrayTruffleObject(int[] array) {
+        ArrayTruffleObject(Object[] array) {
             this.array = array;
         }