+++ graal-js/src/com.oracle.truffle.js.test/src/com/oracle/truffle/js/test/polyglot/JavaScriptLanguageTest.java
@@ -113,9 +113,9 @@ public class JavaScriptLanguageTest {
             Value result = context.eval(JavaScriptLanguage.ID, "[array];");
             assertTrue(result.toString(), result.toString().contains("0, 1, 2, 3, 4"));
             result = context.eval(JavaScriptLanguage.ID, "String([array]);");
-            assertTrue(result.toString(), result.asString().contains("0, 1, 2, 3, 4"));
+            assertTrue(result.toString(), result.asString().contains("0,1,2,3,4"));
             result = context.eval(JavaScriptLanguage.ID, "'' + array;");
-            assertTrue(result.toString(), result.asString().contains("0, 1, 2, 3, 4"));
+            assertTrue(result.toString(), result.asString().contains("0,1,2,3,4"));
         }
     }
 
@@ -126,9 +126,9 @@ public class JavaScriptLanguageTest {
             Value result = context.eval(JavaScriptLanguage.ID, "[obj];");
             assertTrue(result.toString(), result.toString().contains("{answer: 42}"));
             result = context.eval(JavaScriptLanguage.ID, "String([obj]);");
-            assertTrue(result.toString(), result.asString().contains("{answer: 42}"));
+            assertTrue(result.toString(), result.asString().contains("[object Object]"));
             result = context.eval(JavaScriptLanguage.ID, "'' + obj;");
-            assertTrue(result.toString(), result.asString().contains("{answer: 42}"));
+            assertTrue(result.toString(), result.asString().contains("[object Object]"));
         }
     }
 
@@ -137,6 +137,27 @@ public class JavaScriptLanguageTest {
         try (Context context = JSTest.newContextBuilder().build()) {
             Value result = context.eval(JavaScriptLanguage.ID, "[1,[2,[3,[4,[5]]]]];");
             assertEquals("(2)[1, [2, [3, Array(2)]]]", result.toString());
+        }
+    }
+
+    @Test
+    public void testToPrimitiveHostObject() {
+        try (Context context = Context.newBuilder(JavaScriptLanguage.ID).allowAllAccess(true).build()) {
+            context.getBindings(JavaScriptLanguage.ID).putMember("obj", new TestHostObject());
+            Value res = context.eval(JavaScriptLanguage.ID, "obj + obj");
+            assertEquals(84, res.asInt());
+        }
+    }
+
+    public static final class TestHostObject {
+        @SuppressWarnings("static-method")
+        public int valueOf() {
+            return 42;
+        }
+
+        @Override
+        public String toString() {
+            return "string";
         }
     }
 }